ca_ods database project update to match db

+++ ca_ods/dbo/Tables/ref_lookup_county_region.sql
@@ -2,9 +2,11 @@
     [county_cd]     INT           NULL,
     [county]        VARCHAR (255) NULL,
     [region_cd]     INT           NULL,
-    [old_region_cd] FLOAT (53)    NULL,
+    [old_region_cd] INT           NULL,
     [region_6_cd]   SMALLINT      NULL,
     [region_6_tx]   VARCHAR (255) NULL,
     [small_fl]      SMALLINT      NULL
 );
+
+
